Make Presentation adapter generic so it can be reused
